requirements.txt
requests>=2.27.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
yfinance==0.2.36
ta>=0.10.0
matplotlib>=3.5.0
seaborn==0.13.1
prophet>=1.1.1
fastapi==0.109.1
uvicorn==0.27.0
streamlit>=1.20.0
newsapi-python==0.2.7
backtrader==1.9.78.123
sqlalchemy==2.0.25
python-dotenv>=0.20.0
plotly>=5.11.0
optuna>=2.10.0
lightgbm>=3.3.0
joblib>=1.1.0
TA-Lib>=0.4.0
pystan>=2.19.1.1
holidays<2.0.0 